test(node18-ipv6): change default host from localhost to 127.0.0.1
Starting in Node v18, Node resolves an address (like `localhost`)
by returning IP addresses in the order they are returned from the
OS resolver/DNS. [0] This means that on default-IPv6 systems like
a GitHub Actions runner, the link-local address `::1` is returned
instead of `127.0.0.1`.

Unfortunately this new behavior causes a bug in `@open-draft/test-server`,
which serializes an IPv6 host into an invalid URL, because it
incorrectly does not surround the host with square brackets
like it should (e.g. `http://[::1]:8181`).

Since I could not locate the source of `@open-draft/test-server`
to fix this bug, the work-around is to replace all references
to `localhost` with `127.0.0.1` directly - AFAICT there is no
good reason to test this resolution behavior anyway.

[0] https://github.com/nodejs/node/issues/40537