build: upgrade jest and related dependencies
Upgrade to Jest 28, which is necessary because earlier versions
do not export the `fetch` global (among others). In Jest 28
the system for injecting globals is changed to fix this issue.

See: https://jestjs.io/blog/2022/04/25/jest-28#all-nodejs-globals

Also upgrade related dependencies and update import path
of `AsymmetricMatchers` which was moved in this version.