fix(fetch): only set base URL to `location.origin` when `location` is defined
Now that Node v17+ has "native fetch", it's possible to have
`globalThis.fetch` but not `location`, both in our own test
code and in user test code.

So, if `location` is not defined, like in `node-native.fetch` integration
tests, then do not try to set it as the base url of a new URL -
the caller is responsible for giving a correct location.