test(node-native): add initial smoke tests for node-native fetch interception
Only `http` tests pass, and this is only the one test file (still
needs `.body`, `.clone` and `.request` variants).

The tests are copied from the general `fetch.test.ts` file,
which is similar because it also runs in a Node environment,
but different because it uses `cross-fetch`, while this new
file uses native `fetch` (note there is no import of a polyfill).