build(deps-dev): add `undici` as dev dependency to get `Agent`
Add `undici` as a development dependency, because we need to
import `undici.Agent` so that we can set a dispatcher that
intercepts insecure https (equivalent of `httpsAgent` with
previous versions of Node).

Technically we could accomplish this without a dependency
on Undici, but it would require re-implementing the entire
`undici.Agent` class, which would be silly.

Note this is a _development_ dependency. It is not a regular
dependency, because we aren't adding any special undici interceptor;
we're only using the `Agent` because it's the only choice we
have (Node does not re-export its undici module; we must depend on it)